Update MainActivity.java
The code as-is was changing the "Your Requests" header to the name of the request rather than the box? Not sure if this was intentional, but I was assuming we would have sections, tasks and requests, and multiple potential boxes in each (in case you want to do multiple things). So making a new request would effectively create these boxes, completing them would erase them. Since we're short on time, we can work with just the two boxes for now, but it would be nice to keep the headers for usability.